[~] move comments controller to initiatives/
